+++ Estimator.Web/Models/Estimate.cs
@@ -6,9 +6,9 @@ namespace Estimator.Web.Models
     public class Estimate
     {
         public Guid Id {get; set;}
-        public int OptimisticEstimate { get; set; }
-        public int MostLikelyEstimate {get;set;}
-        public int PessimisticEstimate { get; set; }
+        public decimal OptimisticEstimate { get; set; }
+        public decimal MostLikelyEstimate {get;set;}
+        public decimal PessimisticEstimate { get; set; }
         public Customer Customer {get ; set;}
         public IEnumerable<Metric> Metrics {get ; set;}
     }
